% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subtype.R
\name{create_subtype}
\alias{create_subtype}
\title{Utilize 10 clustering machine learning algorithms for cluster analysis}
\usage{
create_subtype(
  data = data.list,
  clust.num = opt.res$N.clust,
  methodslist = list("IntNMF", "CIMLR", "PINSPlus", "NEMO", "COCA", "MoCluster",
    "LRAcluster", "iClusterBayes", "SNF", "ConsensusClustering")
)
}
\arguments{
\item{data}{Matrix file containing two sets of genes. .}

\item{clust.num}{Number of clusters.}
}
\value{
A list of results returned by each specified algorithms.
}
\description{
Utilize 10 clustering machine learning algorithms for cluster analysis
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Pierre-Jean M, Deleuze J F, Le Floch E, et al. Clustering and variable selection evaluation of 13 unsupervised methods for multi-omics data integration\link{J}. Briefings in Bioinformatics, 2019.

intNMF:
Chalise P, Fridley BL. Integrative clustering of multi-level omic data based on non-negative matrix factorization algorithm. PLoS One. 2017;12(5):e0176278.

CIMLR:
Ramazzotti D, Lal A, Wang B, Batzoglou S, Sidow A. Multi-omic tumor data reveal diversity of molecular mechanisms that correlate with survival. Nat Commun. 2018;9(1):4453.

PINSPlus:
Nguyen H, Shrestha S, Draghici S, Nguyen T. PINSPlus: a tool for tumor subtype discovery in integrated genomic data. Bioinformatics. 2019;35(16):2843-2846.

NEMO:
Rappoport N, Shamir R. NEMO: cancer subtyping by integration of partial multi-omic data. Bioinformatics. 2019;35(18):3348-3356.

COCA:
Hoadley KA, Yau C, Wolf DM, et al. Multiplatform analysis of 12 cancer types reveals molecular classification within and across tissues of origin. Cell. 2014;158(4):929-944.

Mocluster:
Meng C, Helm D, Frejno M, Kuster B. moCluster: Identifying Joint Patterns Across Multiple Omics Data Sets. J Proteome Res. 2016;15(3):755-765.

LRAcluster:
Wu D, Wang D, Zhang MQ, Gu J. Fast dimension reduction and integrative clustering of multi-omics data using low-rank approximation: application to cancer molecular classification. BMC Genomics. 2015;16:1022.

iClusterBayes:
Mo Q, Shen R, Guo C, Vannucci M, Chan KS, Hilsenbeck SG. A fully Bayesian latent variable model for integrative clustering analysis of multi-type omics data. Biostatistics. 2018;19(1):71-86.

SNF:
Wang B, Mezlini AM, Demir F, et al. Similarity network fusion for aggregating data types on a genomic scale. Nat Methods. 2014;11(3):333-337.

ConsensusClustering:
Monti S, Tamayo P, Mesirov J, et al. Consensus Clustering: A Resampling-Based Method for Class Discovery and Visualization of Gene Expression Microarray Data. Machine Learning. 2003;52:91-118.
}
