% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_subtype.R
\name{evaluate_subtype}
\alias{evaluate_subtype}
\title{评估亚型}
\usage{
evaluate_subtype(
  expr = "",
  genelist1.file = "",
  genelist2.file = "",
  genelist1.name = "",
  genelist2.name = "",
  clust.num = 2:8,
  center = TRUE,
  scale = TRUE,
  filepath = getwd(),
  filename = "optimal_number_cluster"
)
}
\arguments{
\item{expr}{包含表达数据的文件路径}

\item{clust.num}{聚类数目}

\item{genelist1}{包含基因列表1的文件路径}

\item{genelist2}{包含基因列表2的文件路径}
}
\value{
包含每个预测模型的AUC的数据框
}
\description{
通过开发类别预测模型后计算曲线下面积
}
\examples{
\dontrun{
evaluate_subtype(expr = "expression_data.csv",
               genelist1 = "genelist1.csv",
               genelist2 = "genelist2.csv",
               clust.num = 2:8,
               center = TRUE,
               scale = TRUE,
               filepath = getwd(),
               filename ='optimal_number_cluster')
}

}
