% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_merge.R
\name{filter_merge}
\alias{filter_merge}
\title{Calculating the area under the curve after developing the category predictive model}
\usage{
filter_merge(
  DEGs1 = data1,
  DEGs2 = data2,
  genelist = genes,
  color = c("#CE5C69", "#5770A6", "#E0A980"),
  labels = c("DEGs of GSE6752", "DEGs of GSE48403 ", "Disulfidptosis-related genes"),
  save.path = "G:/importance/undergraduated/12_R包开发/ctDNA/VENE.TIF"
)
}
\arguments{
\item{res.by.ML.Dev.Pred.Category.Sig}{Output of function ML.Dev.Pred.Category.Sig}

\item{cohort.for.cal}{A data frame with the 'ID' and 'Var' as the first two columns. Starting in the fourth column are the variables that contain variables of the model you want to build. The second column 'Var' only contains 'Y' or 'N'.}
}
\value{
A data frame containing the AUC of each predictive model.
}
\description{
Calculating the area under the curve after developing the category predictive model
}
