% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtype_harmony.R
\name{subtype_harmony}
\alias{subtype_harmony}
\title{Get subtypes from consensus clustering of integrative clustering algorithms}
\usage{
subtype_harmony(
  clust.res.list = clust.res.list,
  distance = "euclidean",
  linkage = "average",
  map.col = c("#364888", "#5770A6", "#A281B1", "#E0A980", "#F0A8B9", "#CE5C69",
    "#b30c2a"),
  clust.col = c("#5770A6", "#A281B1", "#CE5C69", "#BDD5EA", "#FFA5AB", "#011627",
    "#023E8A", "#9D4EDD"),
  showID = FALSE,
  width = 5,
  height = 5.5,
  file.save.path = getwd(),
  file.name = "Consensusheatmap"
)
}
\arguments{
\item{clust.res.list}{A list of object returned by create_subtype}

\item{distance}{A string value of distance measurement for hierarchical clustering; 'euclidean' by default.}

\item{linkage}{A string value of clustering method for hierarchical clustering; 'ward.D' by default.}

\item{map.col}{A string vector for heatmap mapping color.}

\item{clust.col}{A string vector storing colors for annotating each cluster at the top of heatmap.}

\item{showID}{A logic value to indicate if showing the sample ID; FALSE by default.}

\item{width}{A numeric value to indicate the width of output figure.}

\item{height}{A numeric value to indicate the height of output figure.}

\item{file.save.path}{A string value to indicate the output path for storing the consensus heatmap.}

\item{file.name}{A string value to indicate the name of the consensus heatmap.}
}
\value{
A consensus heatmap  and a list contains the following components:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    \code{hm} an object returned by \link[ComplexHeatmap]\{pheatmap\}

    \code{matrix}   a similary matrix for pair-wise samples with entries ranging from 0 to 1

    \code{sil}          a silhouette object that can be further passed to \link[MOVICS]\{getSilhouette\}

    \code{clust.res}    a data.frame storing sample ID and corresponding clusters

    \code{clust.dend}   a dendrogram of sample clustering

    \code{method}    a string value indicating the method used for integrative clustering
}\if{html}{\out{</div>}}
}
\description{
Since this R package integrates 10 mainstream clustering algorithms, we borrow the idea of consensus clustering for later integration of the clustering results derived from different algorithms, so as to improve the clustering robustness. The simplest way to run \code{getConsensusMOIC()} is to pass a list of object returned by \verb{get\\\%algorithm_name\\\%()} or by \code{getMOIC()} with specific argument of \code{methodslist}.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Lu, X., Meng, J., Zhou, Y., Jiang, L., and Yan, F. (2020). MOVICS: an R package for multi-omics integration and visualization in cancer subtyping. Bioinformatics, btaa1018. \link{doi.org/10.1093/bioinformatics/btaa1018}

Gu Z, Eils R, Schlesner M (2016). Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics, 32(18):2847-2849.
}
