% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtype_heatmap.R
\name{subtype_heatmap}
\alias{subtype_heatmap}
\title{Get multi-omics comprehensive heatmap}
\usage{
subtype_heatmap(
  data = data.list,
  method = "iClusterBayes",
  genelist1.name = "Disulfidptosis",
  genelist2.name = "Glucose_starvation",
  feat.num = 10,
  col1 = c("#00FF00", "#008000", "#000000", "#800000", "#FF0000"),
  col2 = c("#5770A6", "white", "#CE5C69"),
  clust.col = c("#5770A6", "#A281B1", "#CE5C69", "#BDD5EA", "#FFA5AB", "#011627",
    "#023E8A", "#9D4EDD"),
  genelist.title = c("Disulfidptosis", "Glucose starvation"),
  legend.name = c("mRNA", "mRNA "),
  clust.dend = NULL,
  show.rownames = c(F, F),
  show.colnames = FALSE,
  annCol = NULL,
  annColors = NULL,
  width = 10,
  height = 5,
  file.save.path = getwd(),
  file.name = "subtype heatmap"
)
}
\arguments{
\item{data}{A list of data frame or matrix storing multiple omics data with rows for features and columns for samples.}

\item{method}{A string to select each methon.}

\item{clust.col}{A string vector storing colors for annotating each subtype at the top of heatmap.}

\item{genelist.title}{A string vector to assign titles for each geneset.}

\item{legend.name}{A string vector to assign legend title for each subdata.}

\item{clust.dend}{A dendrogram object returned returned by \code{getMOIC()} with one specified algorithm or \verb{get\\\%algorithm_name\\\%} or \code{getConsensusMOIC()} with a list of multiple algorithms.}

\item{show.rownames}{A logical vector to indicate if showing the names for row of each subdata.}

\item{show.colnames}{A logical vector to indicate if showing the names for column at the bottom of heatmap.}

\item{annCol}{A data.frame storing annotation information for samples with exact the same sample order with data parameter.}

\item{annColors}{A list of string vectors for colors matched with annCol.}

\item{width}{An integer value to indicate the width for each subheatmap with unit of cm.}

\item{height}{An integer value to indicate the height for each subheatmap with unit of cm.}

\item{file.save.path}{A string value to indicate the output path for storing the comprehensive heatmap.}

\item{file.name}{A string value to indicate the name of the comprehensive heatmap.}

\item{show.col.dend}{A logical vector to indicate if showing the dendrogram for column at the top of heatmap.}

\item{show.row.dend}{A logical vector to indicate if showing the dendrogram for row of each subdata.}

\item{clust.dist.row}{A string vector to assign distance method for clustering each subdata at feature dimension.}

\item{clust.method.row}{A string vector to assign clustering method for clustering each subdata at feature dimension.}

\item{color}{A list of string vectors storing colors for each subheatmap of subdata.}
}
\value{
A pdf of multi-omics comprehensive heatmap
}
\description{
This function vertically concatenates multiple heatmap derived from each omics data. \code{subtype_heatmap} supports customized column annotation and is able to mark the selected features if indicated.
}
\examples{
# There is no example and please refer to vignette.
}
\references{
Lu, X., Meng, J., Zhou, Y., Jiang, L., and Yan, F. (2020). MOVICS: an R package for multi-omics integration and visualization in cancer subtyping. Bioinformatics, btaa1018. \link{doi.org/10.1093/bioinformatics/btaa1018}

Gu Z, Eils R, Schlesner M (2016). Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics.
}
