% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtype_KM.R
\name{subtype_KM}
\alias{subtype_KM}
\title{Comparison of prognosis by Kaplan-Meier survival curve}
\usage{
subtype_KM(
  clust.res.list = harmony.res,
  method = "all",
  cli.data = cli.data,
  subtype.list = c("CS1", "CS3"),
  cli.limit = NULL,
  break.time.by = 12,
  clust.col = c("#2EC4B6", "#E71D36", "#FF9F1C", "#BDD5EA", "#FFA5AB"),
  p.adjust.method = "BH",
  surv.median.line = "none",
  file.save.path = getwd(),
  file.name = NULL
)
}
\arguments{
\item{clust.res.list}{An object returned by \code{subtype_harmony()}}

\item{cli.data}{A data.frame with rownames of observations and with at least two columns of \code{time} for survival time and \code{state} for survival status (0: censoring; 1: event)}

\item{clust.col}{A string vector storing colors for each subtype.}

\item{p.adjust.method}{A string value for indicating method for adjusting p values (see \link[stats]{p.adjust}). Allowed values include one of c(\code{holm}, \code{hochberg}, \code{hommel}, \code{bonferroni}, \code{BH}, \code{BY}, \code{fdr}, \code{none}); "BH" by default.}

\item{surv.median.line}{A string value for drawing a horizontal/vertical line at median survival. Allowed values include one of c(\code{none}, \code{hv}, \code{h}, \code{v}). v: vertical, h:horizontal; "none" by default.}

\item{file.save.path}{A string value to indicate the name of the kaplan-meier curve.}

\item{file.name}{A string value to indicate the output path for storing the kaplan-meier curve.}
}
\value{
A figure of multi-omics Kaplan-Meier curve (.pdf) and a list with the following components:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    \code{fitd}       an object returned by \link[survival]\{survdiff\}.

    \code{fid}        an object returned by \link[survival]\{survfit\}.

    \code{overall.p}  a nominal p.value calculated by Kaplain-Meier estimator with log-rank test.

    \code{pairwise.p} an object of class "pairwise.htest" which is a list containing the p values (see \link[survminer]\{pairwise_survdiff\}); (\\emph\{only returned when more than 2 subtypes are identified\}).
}\if{html}{\out{</div>}}
}
\description{
This function calculates Kaplan-meier estimator and generate survival curves with log-rank test to detect prognostic difference among identified subtypes. If more than 2 subtypes are identified, pair-wise comparisons will be performed with an additional table printed on the survival curve.
}
\examples{
# There is no example and please refer to vignette.
}
